Rename last-filter to LastFilter
